pyproject.toml
[tool.poetry]
name = "diario-do-clima-backend"
version = "0.0.1"
description = "Diario do clima backend app"
authors = ["Paulo Roberto Cruz <paulo.cruz9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
python-decouple = "^3.6"
psycopg2-binary = "^2.9.3"
dj-database-url = "^0.5.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.1.0"
requests = "^2.27.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
